Compact native provenance and remove redundant Android wakeups
