+++ .gitignore
@@ -123,5 +123,6 @@
 **/bin/
 **/obj/
 **/artifacts/
+**/__pycache__/
 *.user
 *.lscache
